rewrote individual tests
