--- src/test/java/org/braekpo1nt/mctmanager/games/game/spleef/config/SpleefStorageUtilTest2.java
@@ -1,33 +1,0 @@
-package org.braekpo1nt.mctmanager.games.game.spleef.config;
-
-import be.seeseemelk.mockbukkit.MockBukkit;
-import be.seeseemelk.mockbukkit.ServerMock;
-import org.braekpo1nt.mctmanager.Main;
-import org.braekpo1nt.mctmanager.MyCustomServerMock;
-import org.braekpo1nt.mctmanager.TestUtils;
-import org.braekpo1nt.mctmanager.games.game.config.GameConfigStorageUtil;
-import org.braekpo1nt.mctmanager.games.game.config.GameConfigStorageUtilTest;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-
-import java.io.File;
-import java.io.InputStream;
-import java.util.logging.Level;
-
-public class SpleefStorageUtilTest2 extends GameConfigStorageUtilTest<SpleefConfig> {
-
-    protected SpleefStorageUtilTest2() {
-        super("validSpleefConfig.json", "invalidSpleefConfig.json", "spleefConfig.json");
-    }
-
-    @BeforeEach
-    void setupServerAndPlugin() {
-        ServerMock server = MockBukkit.mock(new MyCustomServerMock());
-        server.getLogger().setLevel(Level.OFF);
-        plugin = MockBukkit.load(Main.class);
-        storageUtil = new SpleefStorageUtil(plugin.getDataFolder());
-    }
-    
-}
